Add flag -short-lived
Kubernetes or docker has an issue where the hostPath does not get updated as the
mounts under the hostPath change on the node. This commit adds a workaround
where after a user specified duration the program will exit. This will cause the
daemonset to restart the pod and have an update mounts in the hostPath.

Workaround for issue #1